Move SchemaIcon.
