+++ src/components/common/SchemaIcon/index.jsx
